split up commands into cogs
